--- alembic/versions/2024_03_23_2233-2024_03_23_2233_comment.py
@@ -1,87 +1,0 @@
-"""COMMENT
-
-Revision ID: 2024_03_23_2233
-Revises: 6ed374791d11
-Create Date: 2024-03-23 22:33:41.137314
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import postgresql
-
-# revision identifiers, used by Alembic.
-revision: str = '2024_03_23_2233'
-down_revision: Union[str, None] = '6ed374791d11'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('participation_status',
-    sa.Column('code', sa.String(), nullable=False),
-    sa.Column('name', sa.String(), nullable=False),
-    sa.Column('to_list', sa.Boolean(), nullable=False),
-    sa.Column('comment', sa.String(), nullable=True),
-    sa.PrimaryKeyConstraint('code', name=op.f('pk_participation_status'))
-    )
-    op.create_table('transport_type',
-    sa.Column('code', sa.String(), nullable=False),
-    sa.Column('name', sa.String(), nullable=False),
-    sa.Column('comment', sa.String(), nullable=True),
-    sa.PrimaryKeyConstraint('code', name=op.f('pk_transport_type'))
-    )
-    op.create_table('participation',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('year', sa.Integer(), nullable=False),
-    sa.Column('person_id', sa.String(), nullable=False),
-    sa.Column('direction_id', sa.String(), nullable=False),
-    sa.Column('role_code', sa.String(), nullable=True),
-    sa.Column('participation_code', sa.String(), nullable=False),
-    sa.Column('status_code', sa.String(), nullable=False),
-    sa.Column('notion_id', sa.String(), nullable=True),
-    sa.ForeignKeyConstraint(['direction_id'], ['direction.notion_id'], name=op.f('fk_participation_direction_id_direction')),
-    sa.ForeignKeyConstraint(['participation_code'], ['participation_type.code'], name=op.f('fk_participation_participation_code_participation_type')),
-    sa.ForeignKeyConstraint(['person_id'], ['person.notion_id'], name=op.f('fk_participation_person_id_person')),
-    sa.ForeignKeyConstraint(['role_code'], ['participation_role.code'], name=op.f('fk_participation_role_code_participation_role')),
-    sa.ForeignKeyConstraint(['status_code'], ['participation_status.code'], name=op.f('fk_participation_status_code_participation_status')),
-    sa.PrimaryKeyConstraint('id', name=op.f('pk_participation'))
-    )
-    op.create_table('arrival',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('badge_id', sa.String(), nullable=False),
-    sa.Column('arrival_date', sa.Date(), nullable=False),
-    sa.Column('arrival_transport', sa.String(), nullable=True),
-    sa.Column('arrival_registered', sa.TIMESTAMP(), nullable=True),
-    sa.Column('departure_date', sa.Date(), nullable=False),
-    sa.Column('departure_transport', sa.String(), nullable=True),
-    sa.Column('departure_registered', sa.TIMESTAMP(), nullable=True),
-    sa.Column('extra_data', postgresql.JSONB(astext_type=sa.Text()), nullable=True),
-    sa.Column('comment', sa.String(), nullable=True),
-    sa.ForeignKeyConstraint(['badge_id'], ['badge.notion_id'], name=op.f('fk_arrival_badge_id_badge')),
-    sa.PrimaryKeyConstraint('id', name=op.f('pk_arrival'))
-    )
-    op.create_table('badge_directions',
-    sa.Column('badge_id', sa.String(), nullable=False),
-    sa.Column('direction_id', sa.String(), nullable=False),
-    sa.ForeignKeyConstraint(['badge_id'], ['badge.notion_id'], name=op.f('fk_badge_directions_badge_id_badge')),
-    sa.ForeignKeyConstraint(['direction_id'], ['direction.notion_id'], name=op.f('fk_badge_directions_direction_id_direction')),
-    sa.PrimaryKeyConstraint('badge_id', 'direction_id', name=op.f('pk_badge_directions'))
-    )
-    op.drop_constraint('fk_badge_direction_id_direction', 'badge', type_='foreignkey')
-    op.drop_column('badge', 'direction_id')
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('badge', sa.Column('direction_id', sa.VARCHAR(), autoincrement=False, nullable=True))
-    op.create_foreign_key('fk_badge_direction_id_direction', 'badge', 'direction', ['direction_id'], ['notion_id'])
-    op.drop_table('badge_directions')
-    op.drop_table('arrival')
-    op.drop_table('participation')
-    op.drop_table('transport_type')
-    op.drop_table('participation_status')
-    # ### end Alembic commands ###
