--- alembic/versions/2024_03_26_1847-2024_03_26_1847_comment.py
@@ -1,30 +1,0 @@
-"""COMMENT
-
-Revision ID: 2024_03_26_1847
-Revises: 2024_03_23_2233
-Create Date: 2024-03-26 18:47:51.074265
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '2024_03_26_1847'
-down_revision: Union[str, None] = '2024_03_23_2233'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
