--- alembic/versions/2024_03_29_1829-2024_03_29_1829_add_primary_key_to_badge_directions.py
@@ -1,36 +1,0 @@
-"""add primary key to badge_directions
-
-Revision ID: 2024_03_29_1829
-Revises: 2024_03_26_1847
-Create Date: 2024-03-29 18:29:51.871539
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '2024_03_29_1829'
-down_revision: Union[str, None] = '2024_03_26_1847'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint('fk_badge_directions_badge_id_badge', 'badge_directions', type_='foreignkey')
-    op.drop_constraint('fk_badge_directions_direction_id_direction', 'badge_directions', type_='foreignkey')
-    op.create_foreign_key(op.f('fk_badge_directions_direction_id_direction'), 'badge_directions', 'direction', ['direction_id'], ['notion_id'], onupdate='CASCADE')
-    op.create_foreign_key(op.f('fk_badge_directions_badge_id_badge'), 'badge_directions', 'badge', ['badge_id'], ['notion_id'], onupdate='CASCADE', ondelete='CASCADE')
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint(op.f('fk_badge_directions_badge_id_badge'), 'badge_directions', type_='foreignkey')
-    op.drop_constraint(op.f('fk_badge_directions_direction_id_direction'), 'badge_directions', type_='foreignkey')
-    op.create_foreign_key('fk_badge_directions_direction_id_direction', 'badge_directions', 'direction', ['direction_id'], ['notion_id'])
-    op.create_foreign_key('fk_badge_directions_badge_id_badge', 'badge_directions', 'badge', ['badge_id'], ['notion_id'])
-    # ### end Alembic commands ###
