+++ alembic/versions/2024_03_30_2140-2024_03_30_2140_badge_add_m2m_connection_direction.py
@@ -1,0 +1,46 @@
+"""badge add M2M connection direction
+
+Revision ID: 2024_03_30_2140
+Revises: 6ed374791d11
+Create Date: 2024-03-30 21:40:33.197691
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = '2024_03_30_2140'
+down_revision: Union[str, None] = '6ed374791d11'
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('badge_directions',
+    sa.Column('badge_id', sa.String(), nullable=False),
+    sa.Column('direction_id', sa.String(), nullable=False),
+    sa.ForeignKeyConstraint(['badge_id'], ['badge.notion_id'], name=op.f('fk_badge_directions_badge_id_badge'), onupdate='CASCADE', ondelete='CASCADE'),
+    sa.ForeignKeyConstraint(['direction_id'], ['direction.notion_id'], name=op.f('fk_badge_directions_direction_id_direction'), onupdate='CASCADE'),
+    sa.PrimaryKeyConstraint('badge_id', 'direction_id', name=op.f('pk_badge_directions'))
+    )
+    op.drop_constraint('fk_badge_direction_id_direction', 'badge', type_='foreignkey')
+    op.drop_column('badge', 'direction_id')
+    op.alter_column('participation', 'notion_id',
+               existing_type=sa.VARCHAR(),
+               nullable=False)
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.alter_column('participation', 'notion_id',
+               existing_type=sa.VARCHAR(),
+               nullable=False)
+    op.add_column('badge', sa.Column('direction_id', sa.VARCHAR(), autoincrement=False, nullable=True))
+    op.create_foreign_key('fk_badge_direction_id_direction', 'badge', 'direction', ['direction_id'], ['notion_id'])
+    op.drop_table('badge_directions')
+    # ### end Alembic commands ###
